# apps/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/refund/wave": {
      "post": {
        "description": "Crée un remboursement via la fonction edge Wave. La forme de la réponse suit le JSON du fournisseur.",
        "operationId": "RefundsController_createWaveRefund",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Résultat du remboursement (corps JSON de la fonction edge)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Entrée invalide"
          },
          "500": {
            "description": "Échec du traitement du remboursement"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Lancer un remboursement Wave",
        "tags": ["Remboursements"]
      }
    },
    "/products": {
      "get": {
        "description": "Renvoie tous les produits de l'organisation du marchand authentifié, avec prix et frais intégrés. Utile pour afficher le catalogue ou intégrer des systèmes externes.",
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Décalage pour la pagination",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre maximal de résultats",
            "schema": {
              "example": 15,
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filtrer par statut actif",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des produits avec prix embarqués",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Lister les produits",
        "tags": ["Produits"]
      },
      "post": {
        "description": "Crée un produit avec un ou plusieurs tarifs en une seule requête. Au moins un prix est requis. Le premier prix, ou celui avec `is_default`, est utilisé lorsque aucun prix n'est précisé au paiement.",
        "operationId": "ProductsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Produit créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Entrée invalide ou erreur de validation"
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Créer un produit",
        "tags": ["Produits"]
      }
    },
    "/products/{id}": {
      "get": {
        "description": "Renvoie un produit avec ses prix et frais. Réponse 404 s'il n'existe pas ou n'est pas accessible avec cette clé API.",
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID du produit",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails du produit avec prix embarqués",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          },
          "404": {
            "description": "Produit introuvable ou accès refusé"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Obtenir un produit par ID",
        "tags": ["Produits"]
      }
    },
    "/products/{id}/prices": {
      "post": {
        "description": "Ajoute une nouvelle option tarifaire à un produit (palier, devise ou période). Un produit ne peut avoir au plus 3 prix actifs. Les prix existants ne se modifient pas : créez-en un nouveau.",
        "operationId": "ProductsController_addPrice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID du produit",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Prix ajouté avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Entrée invalide ou nombre maximal de prix dépassé"
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          },
          "404": {
            "description": "Produit introuvable ou accès refusé"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Ajouter un prix à un produit",
        "tags": ["Produits"]
      }
    },
    "/products/{id}/prices/{priceId}/set-default": {
      "post": {
        "description": "Indique le prix utilisé par défaut pour ce produit (par exemple si le paiement ne fournit pas d'ID de prix). Un seul défaut à la fois ; un nouveau défaut remplace l'ancien.",
        "operationId": "ProductsController_setDefaultPrice",
        "parameters": [
          {
            "name": "priceId",
            "required": true,
            "in": "path",
            "description": "UUID du prix",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID du produit",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prix par défaut mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          },
          "404": {
            "description": "Produit ou prix introuvable"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Définir le prix par défaut",
        "tags": ["Produits"]
      }
    },
...
    "/payout/wave": {
      "post": {
        "description": "Crée un virement via la fonction edge Wave. La forme de la réponse suit le JSON du fournisseur.",
        "operationId": "PayoutsController_createWavePayout",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Virement initié (corps JSON de la fonction edge Wave)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Entrée invalide ou erreur fournisseur"
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Lancer un virement Wave",
        "tags": ["Virements sortants"]
      }
    },
...
    "/webhooks/{id}": {
      "get": {
        "operationId": "WebhooksController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Le webhook",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Obtenir un webhook par ID",
        "tags": ["Webhooks"]
      },
      "patch": {
        "operationId": "WebhooksController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "is_active": {
                    "type": "boolean"
                  },
                  "authorized_events": {
                    "type": "string"
                  },
                  "spi_event_types": {
                    "type": "string"
                  },
                  "supports_spi": {
                    "type": "boolean"
                  },
                  "verification_token": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              },
              "examples": {
                "rotateUrl": {
                  "summary": "Mettre à jour l'URL de livraison",
                  "value": {
                    "url": "https://example.com/webhooks/lomi",
                    "is_active": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook mis à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Mettre à jour un webhook",
        "tags": ["Webhooks"]
      }
    },
...
  },
  "info": {
    "title": "API lomi.",
    "description": "API de traitement des paiements pour les entreprises d'Afrique de l'Ouest francophone.",
    "version": "1.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.lomi.africa",
      "description": "Live"
    },
    {
      "url": "https://sandbox.api.lomi.africa",
      "description": "Test"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    },
    "schemas": {
...
      "RefundResponseDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 123
          },
          "created_at": {
            "type": "string",
            "example": "string"
          },
          "environment": {
            "type": "string",
            "example": "string"
          },
          "fee_amount": {
            "type": "number",
            "example": 123
          },
          "metadata": {
            "type": "object",
            "example": {}
          },
          "reason": {
            "type": "string",
            "example": "string"
          },
          "refund_id": {
            "type": "string",
            "example": "string"
          },
          "refunded_amount": {
            "type": "number",
            "example": 123
          },
          "spi_account_number": {
            "type": "string",
            "example": "string"
          },
          "spi_end2end_id": {
            "type": "string",
            "example": "string"
          },
          "spi_retour_date_demande": {
            "type": "string",
            "example": "string"
          },
          "spi_retour_date_irrevocabilite": {
            "type": "string",
            "example": "string"
          },
          "spi_tx_id": {
            "type": "string",
            "example": "string"
          },
          "status": {
            "type": "string",
            "example": "string"
          },
          "transaction_id": {
            "type": "string",
            "example": "string"
          },
          "updated_at": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "amount",
          "created_at",
          "environment",
          "fee_amount",
          "metadata",
          "reason",
          "refund_id",
          "refunded_amount",
          "spi_account_number",
          "spi_end2end_id",
          "spi_retour_date_demande",
          "spi_retour_date_irrevocabilite",
          "spi_tx_id",
          "status",
          "transaction_id",
          "updated_at"
        ]
      },
      "PriceResponseDto": {
        "type": "object",
        "properties": {
          "price_id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Unique price identifier"
          },
          "amount": {
            "type": "number",
            "example": 10000,
            "description": "Price amount"
          },
          "currency_code": {
            "type": "string",
            "example": "XOF",
            "description": "Currency code",
            "enum": ["XOF", "USD", "EUR"]
          },
          "billing_interval": {
            "type": "string",
            "example": "month",
            "description": "Billing interval for recurring products",
            "enum": ["day", "week", "month", "year"],
            "nullable": true
          },
          "pricing_model": {
            "type": "string",
            "example": "standard",
            "description": "Pricing model",
            "enum": ["standard", "pay_what_you_want", "tiered"]
          },
          "minimum_amount": {
            "type": "number",
            "example": 5000,
            "description": "Minimum amount (for pay_what_you_want)",
            "nullable": true
          },
          "maximum_amount": {
            "type": "number",
            "example": 50000,
            "description": "Maximum amount (for pay_what_you_want)",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "example": true,
            "description": "Whether this price is active"
          },
          "is_default": {
            "type": "boolean",
            "example": true,
            "description": "Whether this is the default price"
          },
          "metadata": {
            "type": "object",
            "example": {
              "notes": "Early bird pricing"
            },
            "description": "Additional metadata",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "example": "2024-01-15T10:30:00Z",
            "description": "When the price was created"
          },
          "updated_at": {
            "type": "string",
            "example": "2024-01-15T10:30:00Z",
            "description": "When the price was last updated"
          }
        },
        "required": [
          "price_id",
          "amount",
          "currency_code",
          "billing_interval",
          "pricing_model",
          "minimum_amount",
          "maximum_amount",
          "is_active",
          "is_default",
          "metadata",
          "created_at",
          "updated_at"
        ]
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Unique product identifier"
          },
          "organization_id": {
            "type": "string",
            "example": "789e0123-e89b-12d3-a456-426614174000",
            "description": "Organization ID"
          },
          "name": {
            "type": "string",
            "example": "Premium Subscription",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "example": "Access to all premium features",
            "description": "Product description",
            "nullable": true
          },
          "product_type": {
            "type": "string",
            "example": "recurring",
            "description": "Product type",
            "enum": ["one_time", "recurring", "usage_based"]
          },
          "images": {
            "example": ["https://example.com/image.png"],
            "description": "Product images URLs",
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "is_active": {
            "type": "boolean",
            "example": true,
            "description": "Whether the product is active"
          },
          "display_on_storefront": {
            "type": "boolean",
            "example": true,
            "description": "Whether to display on storefront"
          },
          "metadata": {
            "type": "object",
            "example": {
              "category": "subscription"
            },
            "description": "Additional metadata",
            "nullable": true
          },
          "failed_payment_action": {
            "type": "string",
            "example": "pause",
            "description": "Action to take on failed payment",
            "enum": ["pause", "cancel", "retry"],
            "nullable": true
          },
          "charge_day": {
            "type": "number",
            "example": 1,
            "description": "Day of month to charge (1-31)",
            "nullable": true
          },
          "first_payment_type": {
            "type": "string",
            "example": "initial",
            "description": "When to charge first payment",
            "enum": ["initial", "non_initial", "prorated"],
            "nullable": true
          },
          "trial_enabled": {
            "type": "boolean",
            "example": false,
            "description": "Whether trial is enabled"
          },
          "trial_period_days": {
            "type": "number",
            "example": 14,
            "description": "Trial period in days",
            "nullable": true
          },
          "usage_aggregation": {
            "type": "string",
            "example": "sum",
            "description": "Usage aggregation method",
            "enum": ["sum", "max", "last_during_period", "last_ever"],
            "nullable": true
          },
          "usage_unit": {
            "type": "string",
            "example": "api_calls",
            "description": "Unit of usage measurement",
            "nullable": true
          },
          "environment": {
            "type": "string",
            "example": "live",
            "description": "Environment",
            "enum": ["test", "live"]
          },
          "prices": {
            "description": "Product prices",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriceResponseDto"
            }
          },
          "fees": {
            "example": [],
            "description": "Associated fees",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "created_at": {
            "type": "string",
            "example": "2024-01-15T10:30:00Z",
            "description": "When the product was created"
          },
          "updated_at": {
            "type": "string",
            "example": "2024-01-15T10:30:00Z",
            "description": "When the product was last updated"
          }
        },
        "required": [
          "product_id",
          "organization_id",
          "name",
          "description",
          "product_type",
          "images",
          "is_active",
          "display_on_storefront",
          "metadata",
          "failed_payment_action",
          "charge_day",
          "first_payment_type",
          "trial_enabled",
          "trial_period_days",
          "usage_aggregation",
          "usage_unit",
          "environment",
          "prices",
          "fees",
          "created_at",
          "updated_at"
        ]
      },
...
      "PayoutResponseDto": {
        "type": "object",
        "properties": {
          "account_id": {
            "type": "string",
            "example": "string"
          },
          "amount": {
            "type": "number",
            "example": 123
          },
          "created_at": {
            "type": "string",
            "example": "string"
          },
          "created_by": {
            "type": "string",
            "example": "string"
          },
          "currency_code": {
            "type": "string",
            "example": "string"
          },
          "environment": {
            "type": "string",
            "example": "string"
          },
          "metadata": {
            "type": "object",
            "example": {}
          },
          "organization_id": {
            "type": "string",
            "example": "string"
          },
          "payout_id": {
            "type": "string",
            "example": "string"
          },
          "payout_method_id": {
            "type": "string",
            "example": "string"
          },
          "provider_code": {
            "type": "string",
            "example": "string"
          },
          "status": {
            "type": "string",
            "example": "string"
          },
          "updated_at": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "account_id",
          "amount",
          "created_at",
          "created_by",
          "currency_code",
          "environment",
          "metadata",
          "organization_id",
          "payout_id",
          "payout_method_id",
          "provider_code",
          "status",
          "updated_at"
        ]
      },
...
    }
  }
}